feat: 조회부분을 QueryService로 변경
